refactor(profile): move `@top/@tracks` to `/views`
