--- app/profile/[id]/@top/layout.tsx
@@ -1,21 +1,0 @@
-'use client'
-
-import type { ReactNode } from 'react'
-
-import type { ProfileLayoutBaseProps } from '@app/profile/types'
-
-namespace ProfileTopGenresSubLayout {
-  export type Props = Readonly<
-    ProfileLayoutBaseProps & {
-      tracks: ReactNode
-    }
-  >
-}
-
-function ProfileTopGenresSubLayout({
-  tracks,
-}: ProfileTopGenresSubLayout.Props) {
-  return tracks
-}
-
-export default ProfileTopGenresSubLayout
