+++ app/profile/[id]/views/index.ts
@@ -1,3 +1,4 @@
 export * from './top-genres-view'
 export * from './top-artists-view'
 export * from './top-albums-view'
+export * from './top-tracks-view'
